코드 리뷰 반영(3)
1. 실수형 연산 가능하도록 수정
검증 단계에서 실수 검증을 할 수 있도록  Ineteger.parseInt()에서 Double.parseDouble()로 변경하였습니다. 연산과 관련된 반환 타입 역시 double로 변경하였습니다.
단 계산 결과 저장 또는 출력시 소수점 첫째자리까지만 표현하도록 하였습니다.

+++ src/main/java/com/calculator/java/Validator.java
@@ -39,7 +39,7 @@ public class Validator {
 
     private boolean isNumber(String element) {
         try {
-            Integer.parseInt(element);
+            Double.parseDouble(element);
             return true;
         } catch (NumberFormatException numberFormatException) {
             return false;
